Move IToken and IProvider::getToken to OCP
