+++ lib/private/Authentication/Token/IToken.php
@@ -26,109 +26,7 @@ declare(strict_types=1);
  */
 namespace OC\Authentication\Token;
 
-use JsonSerializable;
+use OCP\Authentication\Token\IToken as OCPIToken;
 
-interface IToken extends JsonSerializable {
-	public const TEMPORARY_TOKEN = 0;
-	public const PERMANENT_TOKEN = 1;
-	public const WIPE_TOKEN = 2;
-	public const DO_NOT_REMEMBER = 0;
-	public const REMEMBER = 1;
-
-	/**
-	 * Get the token ID
-	 *
-	 * @return int
-	 */
-	public function getId(): int;
-
-	/**
-	 * Get the user UID
-	 *
-	 * @return string
-	 */
-	public function getUID(): string;
-
-	/**
-	 * Get the login name used when generating the token
-	 *
-	 * @return string
-	 */
-	public function getLoginName(): string;
-
-	/**
-	 * Get the (encrypted) login password
-	 *
-	 * @return string|null
-	 */
-	public function getPassword();
-
-	/**
-	 * Get the timestamp of the last password check
-	 *
-	 * @return int
-	 */
-	public function getLastCheck(): int;
-
-	/**
-	 * Set the timestamp of the last password check
-	 *
-	 * @param int $time
-	 */
-	public function setLastCheck(int $time);
-
-	/**
-	 * Get the authentication scope for this token
-	 *
-	 * @return string
-	 */
-	public function getScope(): string;
-
-	/**
-	 * Get the authentication scope for this token
-	 *
-	 * @return array
-	 */
-	public function getScopeAsArray(): array;
-
-	/**
-	 * Set the authentication scope for this token
-	 *
-	 * @param array $scope
-	 */
-	public function setScope($scope);
-
-	/**
-	 * Get the name of the token
-	 * @return string
-	 */
-	public function getName(): string;
-
-	/**
-	 * Get the remember state of the token
-	 *
-	 * @return int
-	 */
-	public function getRemember(): int;
-
-	/**
-	 * Set the token
-	 *
-	 * @param string $token
-	 */
-	public function setToken(string $token);
-
-	/**
-	 * Set the password
-	 *
-	 * @param string $password
-	 */
-	public function setPassword(string $password);
-
-	/**
-	 * Set the expiration time of the token
-	 *
-	 * @param int|null $expires
-	 */
-	public function setExpires($expires);
+interface IToken extends OCPIToken {
 }
