+++ lib/public/Authentication/Token/IProvider.php
@@ -38,4 +38,15 @@ interface IProvider {
 	 * @return void
 	 */
 	public function invalidateTokensOfUser(string $uid, ?string $clientName);
+
+	/**
+	 * Get a token by token string id
+	 *
+	 * @since 28.0.0
+	 * @throws InvalidTokenException
+	 * @throws ExpiredTokenException
+	 * @throws WipeTokenException
+	 * @return IToken
+	 */
+	public function getToken(string $tokenId): IToken;
 }
